+++ artiq/examples/kc705_nist_clock/device_db.py
@@ -17,20 +17,20 @@ device_db = {
     },
     "core_log": {
         "type": "controller",
-        "host": "::1",
+        "host": "localhost",
         "port": 1068,
         "command": "aqctl_corelog -p {port} --bind {bind} " + core_addr
     },
     "core_moninj": {
         "type": "controller",
-        "host": "::1",
+        "host": "localhost",
         "port_proxy": 1383,
         "port": 1384,
         "command": "aqctl_moninj_proxy --port-proxy {port_proxy} --port-control {port} --bind {bind} " + core_addr
     },
     "core_analyzer": {
         "type": "controller",
-        "host": "::1",
+        "host": "localhost",
         "port_proxy": 1385,
         "port": 1386,
         "command": "aqctl_coreanalyzer_proxy --port-proxy {port_proxy} --port-control {port} --bind {bind} " + core_addr