+++ RELEASE_NOTES.rst
@@ -52,6 +52,7 @@ ARTIQ-9 (Unreleased)
 * The Zadig driver installer was added to the MSYS2 offline installer.
 * ``artiq.coredevice.fmcdio_vhdci_eem`` has been removed.
 * Optional SSL support with mutual certificate authentication was added between components running on computers.
+* Default support for systems without IPv6.
 
 Breaking changes: